Move Paper signed arguments impl to own project
