+++ cloud-paper-signed-arguments/src/main/java/org/incendo/cloud/paper/signed/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Paper implementation of signed arguments.
+ */
+package org.incendo.cloud.paper.signed;
