+++ cloud-paper-signed-arguments/src/main/resources/META-INF/services/org.incendo.cloud.minecraft.signed.SignedStringMapper
@@ -1,0 +1,1 @@
+org.incendo.cloud.paper.signed.PaperSignedMapper
