--- cloud-paper/src/main/resources/META-INF/services/org.incendo.cloud.minecraft.signed.SignedStringMapper
@@ -1,1 +1,0 @@
-org.incendo.cloud.paper.PaperSignedMapper
